Implemented efficient GPS time in nanoseconds. TODO: Improve --writeWaveform mode performrmance

+++ src/scanner/Measurement.h
@@ -62,7 +62,7 @@ public:
 	/**
 	 * @brief Measurement GPS time
 	 */
-	long gpsTime;
+	double gpsTime;
 
 	// ***  CONSTRUCTION / DESTRUCTION  *** //
 	// ************************************ //
